Fixing Images for the Last Time - Project Complete
Images wouldn't show up as there was no resource folder in src (it is needed for it to show up in the build folder when you do clean and build).

What I did to fix this:
- Created Resources under src folder
- Added Images to resources folder
- Did a 'clean and build' 🧹🛠️

After doing this, images should work when the program is run on a different computer.